+++ docs/en/changes/changes.md
@@ -4,10 +4,13 @@
 
 #### OAP Server
 
+* BanyanDB: Support `hot/warm/cold` stages configuration.
+
 #### UI
 
 #### Documentation
 
+* BanyanDB: Add `Data Lifecycle Stages(Hot/Warm/Cold)` documentation.
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/230?closed=1)